Avatar totems added to major towns, corviknight network disabled
Added avatar totems in the overworlds of Novo Town, LuxTech Campus, Svait, Prizca West, Prizca East, and Sweetrock Harbor
Moved the Avatar Totem in Whitebloom Town to a more central location.
Moved the Avatar Totem in the Velenz Menagerie somewhat to the west, amongst the willow trees
Cleared much of the grass on the path to Yezera in Whitebloom Town
Made the player respawn at the Whitebloom Town totem if they lose to Yezera there
Disabled the Corviknight Network. Interacting with any of the members of the network now gives a generic interact description (TBD what to do with these).
Removed the Avatar Totem in the Hollowed Layer