Various Mirror Tundra fixes
The Avatar of A. Sandslash now drops the Relic Vase that was previously near by
The Avatar of Froslass now drops the Amulet Coin that was available elsewhere in the Glacial Labyring. The EXP Candy XL that was near the Froslass has been moved to where the Amulet Coin was. Added a rock to make that loot event no longer able to be accessed early from the ledge above.
The Avatar of Hypno now drops 5 disable balls
Recoded how the 3 aspects of Farid disappear after being defeated